Add history, clear, home
I added a cmdline history and changed the cmdline widget to a pulldown to access history. I attempted to make it save partial command strings from where a command failed due to a syntax error, and implemented a syntax check, but that bit is still too primitive to work properly. I also added a Home and Clear button at the top.